Updated with new file on 3

+++ github_committer.py
@@ -34,12 +34,17 @@ def download_xls():
     if not os.path.isdir(put_in_dir):
         os.makedirs(put_in_dir)
 
-    wget.download(DOWNLOAD_LINK)
-    time.sleep(5)
-    try:
-        os.rename(DOWNLOADED_FILE, put_in_dir + DOWNLOADED_FILE)
-        print os.getcwd()
+    if os.path.isfile("rank_1.xls"):
         os.remove("rank_1.xls")
+
+    try:
+        wget.download(DOWNLOAD_LINK)
+        time.sleep(5)
+        os.rename(DOWNLOADED_FILE, put_in_dir + DOWNLOADED_FILE)
+        time.sleep(2)
+        print os.getcwd()
+        if os.path.isfile("rank_1.xls"):
+            os.remove("rank_1.xls")
         log_message += "Downloaded file for today\t"
     except Exception, ex:
         log_message += "Error moving the file: " + str(ex)
